test:  filter-action-row.test.tsx updated all tests
